==> election_tweets/1_ballotpediaScrape.ipynb <==
import pandas as pd
import numpy as np
import re
import time

import requests
from bs4 import BeautifulSoup
from selenium import webdriver

url = 'https://ballotpedia.org/List_of_congressional_candidates_in_the_2022_elections'

state_names = ['Alaska', 'Alabama', 'Arkansas', 'Arizona', 'California', 'Colorado', 'Connecticut', 'Delaware', 'Florida', 'Georgia', 'Hawaii', 'Iowa', 'Idaho', 'Illinois', 'Indiana', 'Kansas', 'Kentucky', 'Louisiana', 'Massachusetts', 'Maryland', 'Maine', 'Michigan', 'Minnesota', 'Missouri', 'Mississippi', 'Montana', 'North Carolina', 'North_Carolina', 'North Dakota', 'North_Dakota', 'Nebraska', 'New Hampshire', 'New_Hampshire', 'New Jersey', 'New_Jersey', 'New Mexico', 'New_Mexico', 'Nevada', 'New York', 'New_York', 'Ohio', 'Oklahoma', 'Oregon', 'Pennsylvania', 'Rhode Island', 'Rhode_Island', 'South Carolina', 'South_Carolina', 'South Dakota', 'South_Dakota', 'Tennessee', 'Texas', 'Utah', 'Virginia', 'Vermont', 'Washington', 'Wisconsin', 'West Virginia', 'West_Virginia', 'Wyoming']
state_regex = '(' + ')|('.join(state_names) + ')'
party_names = ['Democratic Party','Republican Party','Libertarian Party','Independent','Nonpartisan','Unaffiliated','Alaskan Independence Party','American Independent Party','Green Party','No party preference','Peace and Freedom Party of California','American Constitution Party','Independent Party','No Party Affiliation','Constitution Party','Working Class Party','Legal Marijuana Now Party','Working Families Party','Independence Party','Progressive Party','United Utah Party','Vermont Progressive Party','American Solidarity Party','Congress Sucks Party','Logic Party','Alliance Party']
party_regex = '(' + ')|('.join(party_names) + ')'

## launch headless Edge instance
options = webdriver.EdgeOptions()
options.add_argument('--headless')

driver = webdriver.Edge(options = options)
driver.get(url)

## grab html from page
page = driver.page_source.encode('utf-8')
soup = BeautifulSoup(page, 'html.parser')

## loop through states, grab code, and grab links for Senate and House
senate_links = []
house_links = []

for state in state_names:
    state_html = soup.find('div', id = state.replace(' ', '_'))

    senate_html = str(state_html).split('id="U.S._House')[0]
    senate_links.extend(re.findall(r'(?<=href=")https://ballotpedia.org/[A-Za-z\.\'\-\(\)_]+(?=">)', 
                                   str(senate_html))
                       )
    
    house_html = str(state_html).split('id="U.S._House')[1]
    # house_links.extend(re.findall(r'(?<=")(https://ballotpedia.org/([A-Za-z\.\'\-]+_){1,4}[A-Za-z\.\'\-\(\)]+)(?=")',
    house_links.extend(re.findall(r'(?<=href=")https://ballotpedia.org/[A-Za-z\.\'\-\(\)_]+(?=">)',
                                  str(house_html))
                      )

house_links = [x for x in house_links if '_Party' not in x and '/Nonpartisan' not in x]
house_links = [*set(house_links)]

len(house_links)

# function
def scrape_candidate_page(url, type = 'House'):
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    body = soup.find('div', class_ = 'mw-parser-output')
    link_table = soup.find('div', class_ = 'infobox person')
    election_link = body.find('a', href = re.compile(str('https://ballotpedia.org/.*(Congress|House).*,_2022')))

    name = re.search(r'[A-Za-zé\.\'\- ]+(?=\n\n)', link_table.get_text())[0]
    try:
        party = re.search(party_regex, link_table.get_text())[0]
    except:
        party = ''
    try:
        state = re.search(state_regex, str(election_link))[0]
    except:
        state = ''
    try:
        district = re.search(r'(?<=_)[0-9]+(?=[a-z][a-z]_)', str(election_link))[0]
    except:
        district = ''
    try:
        last_election = re.search(r'[A-Za-z]+ [0-9]+, [0-9]+(?=\n\n)', link_table.get_text())[0]
    except:
        last_election = ''
    twitter = re.findall(r'(?<=")https://www.twitter.com/\w+(?=")', str(link_table))
    ballotpedia = url
    
    result = [name, party, state, district, last_election, twitter, ballotpedia]
    
    return result #, body, link_table, election_link;

# result, body, link_table, election_link = scrape_candidate_page(house_links[1967])
result = scrape_candidate_page(house_links[1967])

result

ballotpedia_scrape = pd.DataFrame(columns = ['name', 'party', 'state', 'district', 'last_election', 'twitter', 'ballotpedia'])

for url in house_links:
    result = scrape_candidate_page(url)
    ballotpedia_scrape.loc[len(ballotpedia_scrape)] = result
    # time.sleep(1)
    if house_links.index(url) % 100 == 0:
        print(f'Progress - {house_links.index(url)} of {len(house_links)}')

ballotpedia_scrape

len(ballotpedia_scrape.query('twitter.str.len() == 0'))

ballotpedia_scrape.party.value_counts()

ballotpedia_scrape.query('party == ""')

ballotpedia_scrape.query('state == ""')

# Alaska, Delaware, North Dakota, South Dakota, Vermont, Wyoming
ballotpedia_scrape.query('district == ""').state.unique()

ballotpedia_scrape.district.value_counts()

ballotpedia_scrape.to_csv('data/ballotpedia_scrape2022.csv', index = False)

# url = 'https://ballotpedia.org/List_of_congressional_candidates_in_the_2022_elections'
# page = requests.get(url)
# soup = BeautifulSoup(page.content, 'html.parser')

# html_text = soup.prettify()
# with open('output/ballotpedia_html.txt', 'w', encoding = 'utf-8') as text_file:
#     print(html_text, file = text_file)

# candidate_links = re.findall(r'(?<=")(https://ballotpedia.org/([A-Za-z\.\']+_){1,4}[A-Za-z\.\'\(\)]+)(?=")', str(soup))
# candidate_links = [x[0] for x in candidate_links if '_Party' not in x[0] and 'php' not in x[0] and 'Main_Page' not in x[0]]
# candidate_links = [*set(candidate_links)]

# len(candidate_links)

# state_names = ['Alaska', 'Alabama', 'Arkansas', 'American Samoa', 'Arizona', 'California', 'Colorado', 'Connecticut', 'District of Columbia', 'Delaware', 'Florida', 'Georgia', 'Guam', 'Hawaii', 'Iowa', 'Idaho', 'Illinois', 'Indiana', 'Kansas', 'Kentucky', 'Louisiana', 'Massachusetts', 'Maryland', 'Maine', 'Michigan', 'Minnesota', 'Missouri', 'Mississippi', 'Montana', 'North Carolina', 'North Dakota', 'Nebraska', 'New Hampshire', 'New Jersey', 'New Mexico', 'Nevada', 'New York', 'Ohio', 'Oklahoma', 'Oregon', 'Pennsylvania', 'Puerto Rico', 'Rhode Island', 'South Carolina', 'South Dakota', 'Tennessee', 'Texas', 'Utah', 'Virginia', 'Virgin Islands', 'Vermont', 'Washington', 'Wisconsin', 'West Virginia', 'Wyoming']
# state_regex = '(' + ')|('.join(state_names) + ')'
# party_names = ['Democratic Party','Republican Party','Libertarian Party','Independent','Nonpartisan','Unaffiliated','Alaskan Independence Party','American Independent Party','Green Party','No party preference','Peace and Freedom Party of California','American Constitution Party','Independent Party','No Party Affiliation','Constitution Party','Working Class Party','Legal Marijuana Now Party','Working Families Party','Independence Party','Progressive Party','United Utah Party','Vermont Progressive Party','American Solidarity Party','Congress Sucks Party','Logic Party']
# party_regex = '(' + ')|('.join(party_names) + ')'

# ballotpedia_scrape = pd.DataFrame(columns = ['name', 'party', 'state', 'district', 'last_election', 'twitter', 'ballotpedia'])

# for url in candidate_links[0:101]:
#     page = requests.get(url)
#     soup = BeautifulSoup(page.content, 'html.parser')
#     body = soup.find('div', class_ = 'mw-parser-output')
#     link_table = soup.find('div', class_ = 'infobox person')
#     name = re.search(r'[A-Za-z\.\' ]+(?=\n\n)', link_table.get_text())[0]
#     try:
#         party = re.search(party_regex, link_table.get_text())[0]
#     except:
#         party = ''
#     try:
#         state = re.search(state_regex, body.get_text())[0]
#     except:
#         state = ''
#     try:
#         district = body.find('a', href = re.compile(str('https://ballotpedia.org/' + state + '\'.*,_2022'))).get_text()
#         district = re.search(str('(?<=' + state + '\')(s | )([0-9]+)'), district)[2]
#     except:
#         district = ''
#     try:
#         last_election = re.search(r'[A-Za-z]+ [0-9]+, [0-9]+(?=\n\n)', link_table.get_text())[0]
#     except:
#         last_election = ''
#     twitter = re.findall(r'(?<=")https://www.twitter.com/\w+(?=")', str(link_table))
#     ballotpedia = url
#     ballotpedia_scrape.loc[len(ballotpedia_scrape)] = [name, party, state, district, last_election, twitter, ballotpedia]
#     # time.sleep(1)
#     if candidate_links.index(url) % 100 == 0:
#         print(f'Progress - {candidate_links.index(url)} of {len(candidate_links)}')
